feat: Add Colors to Mesh Structure
The Mesh-Face structure now allows for selecting colors. You can set the color for a mesh, or just for an individual face, during it's initialization. If both are selected, the face takes priority. If no color is selected on the face, it defaults to the mesh's color, and if no mesh color is selected, it defaults to black.